Move password hasher classes into subpackage
